Refactored the onboarding application and wrap it all under its own engine. Also fixed problems with Gemfile & the default template not including the routes
