+++ lib/ecrire/onboarding/routes.rb
@@ -2,9 +2,7 @@ Ecrire::Onboarding::Engine.routes.draw do
 
   root 'onboarding#index'
 
-  namespace :onboarding, only: :index do
-    resources :databases, only: [:index, :create]
-    resources :users, only: [:index, :create]
-  end
+  resources :databases, only: [:index, :create]
+  resources :users, only: [:index, :create]
 end
