+++ lib/ecrire/onboarding/views/databases/index.html.erb
@@ -15,7 +15,7 @@
   </div>
 
 
-  <%= form_tag onboarding_databases_path, method: :post do %>
+  <%= form_tag databases_path, method: :post do %>
 
     <% unless @exception.nil? %>
       <div class='error' as='Message'>
