+++ lib/ecrire/onboarding/views/onboarding/welcome.html.erb
@@ -7,6 +7,6 @@
 
 
 <section class='welcome'>
-  <%= link_to 'Configure', onboarding_databases_path, class: %w(button) %>
+  <%= link_to 'Configure', databases_path, class: %w(button) %>
 </section>
