+++ lib/ecrire/template/views/layouts/application.html.erb
@@ -12,7 +12,7 @@
     <main>
       <%= admin_navigation %>
       <header>
-        <h1><%= link_to 'E', :root %></h1>
+        <h1><%= link_to 'E', '/' %></h1>
         <nav>
           <%= link_to 'Customize', '/customize' %>
           <% if !signed_in? %>
